# WIP: Reduce the number of SQL queries