--- a/src/graphql/schema/collection/createCollectionGqlType.ts
+++ b/src/graphql/schema/collection/createCollectionGqlType.ts
@@ -7,7 +7,6 @@ import createConnectionGqlField from '../connection/createConnectionGqlField'
 import BuildToken from '../BuildToken'
 import createCollectionRelationTailGqlFieldEntries from './createCollectionRelationTailGqlFieldEntries'
 import getConditionGqlType from './getConditionGqlType'
-import { sql } from '../../../postgres/utils'
 
 /**
  * Creates the output object type for a collection. This type will include all
@@ -46,14 +45,13 @@ export default function createCollectionGqlType<TValue> (
       Array.from(type.fields).map(
         <TFieldValue>([fieldName, field]: [string, ObjectType.Field<TValue, TFieldValue>]) => {
           const { gqlType, intoGqlOutput } = getGqlOutputType(buildToken, field.type)
-          const formattedFieldName = formatName.field(fieldName)
           return {
-            key: formattedFieldName,
+            key: formatName.field(fieldName),
             value: {
               description: field.description,
               type: gqlType,
               externalFieldName: fieldName,
-              resolve: (value: TValue): mixed => intoGqlOutput(field.getValue(value),
+              resolve: (value: TValue): mixed => intoGqlOutput(field.getValue(value)),
             },
           }
         },