--- a/src/graphql/schema/collection/createCollectionGqlType.ts
+++ b/src/graphql/schema/collection/createCollectionGqlType.ts
@@ -100,7 +100,9 @@ export default function createCollectionGqlType<TValue> (
                   relation.getTailConditionFromHeadAlias!(aliasIdentifier),
                   conditionFromGqlInput(args.condition),
                 ),
-            }, true),
+              subquery: true,
+              relation,
+            }),
           ]
         }),
     ),