--- a/src/graphql/schema/collection/mutations/createCreateCollectionMutationFieldEntry.ts
+++ b/src/graphql/schema/collection/mutations/createCreateCollectionMutationFieldEntry.ts
@@ -74,7 +74,7 @@ export default function createCreateCollectionMutationFieldEntry <TValue>(
     // When we execute we just create a value in the collection after
     // transforming the correct input field.
     // TODO: test
-    execute: (context, input) =>
-      collection.create!(context, inputFromGqlInput(input[inputFieldName])),
+    execute: (context, input, resolveInfo) =>
+      collection.create!(context, inputFromGqlInput(input[inputFieldName]), resolveInfo),
   })]
 }