--- a/src/graphql/schema/collection/mutations/createCreateCollectionMutationFieldEntry.ts
+++ b/src/graphql/schema/collection/mutations/createCreateCollectionMutationFieldEntry.ts
@@ -27,6 +27,7 @@ export default function createCreateCollectionMutationFieldEntry <TValue>(
 
   return [formatName.field(name), createMutationGqlField<TValue>(buildToken, {
     name,
+    relatedGqlType: collectionGqlType,
     description: `Creates a single ${scrib.type(collectionGqlType)}.`,
 
     inputFields: [