--- a/src/graphql/schema/collection/mutations/createDeleteCollectionKeyMutationFieldEntry.ts
+++ b/src/graphql/schema/collection/mutations/createDeleteCollectionKeyMutationFieldEntry.ts
@@ -5,6 +5,7 @@ import BuildToken from '../../BuildToken'
 import createMutationGqlField from '../../createMutationGqlField'
 import createCollectionKeyInputHelpers from '../createCollectionKeyInputHelpers'
 import { getDeleteCollectionPayloadGqlType } from './createDeleteCollectionMutationFieldEntry'
+import getGqlOutputType from '../../type/getGqlOutputType'
 
 /**
  * Creates a delete mutation which will delete a single value from a collection
@@ -22,6 +23,7 @@ export default function createDeleteCollectionKeyMutationFieldEntry <TValue, TKe
   const { collection } = collectionKey
   const name = `delete-${collection.type.name}-by-${collectionKey.name}`
   const inputHelpers = createCollectionKeyInputHelpers(buildToken, collectionKey)
+  const { gqlType } = getGqlOutputType(buildToken, collection.type)
 
   return [formatName.field(name), createMutationGqlField<TValue>(buildToken, {
     name,
@@ -29,6 +31,6 @@ export default function createDeleteCollectionKeyMutationFieldEntry <TValue, TKe
     inputFields: inputHelpers.fieldEntries,
     payloadType: getDeleteCollectionPayloadGqlType(buildToken, collection),
     execute: (context, input, resolveInfo) =>
-      collectionKey.delete!(context, inputHelpers.getKeyFromInput(input), resolveInfo),
+      collectionKey.delete!(context, inputHelpers.getKeyFromInput(input), resolveInfo, gqlType),
   })]
 }