--- a/src/graphql/schema/collection/mutations/createDeleteCollectionMutationFieldEntry.ts
+++ b/src/graphql/schema/collection/mutations/createDeleteCollectionMutationFieldEntry.ts
@@ -40,13 +40,13 @@ export default function createDeleteCollectionMutationFieldEntry <TValue>(
     payloadType: getDeleteCollectionPayloadGqlType(buildToken, collection),
     // Execute by deserializing the id into its component parts and delete a
     // value in the collection using that key.
-    execute: (context, input) => {
+    execute: (context, input, resolveInfo) => {
       const result = idSerde.deserialize(inventory, input[options.nodeIdFieldName] as string)
 
       if (result.collection !== collection)
         throw new Error(`The provided id is for collection '${result.collection.name}', not the expected collection '${collection.name}'.`)
 
-      return primaryKey.delete!(context, result.keyValue)
+      return primaryKey.delete!(context, result.keyValue, resolveInfo)
     },
   })]
 }