--- a/src/graphql/schema/collection/mutations/createUpdateCollectionKeyMutationFieldEntry.ts
+++ b/src/graphql/schema/collection/mutations/createUpdateCollectionKeyMutationFieldEntry.ts
@@ -5,6 +5,7 @@ import BuildToken from '../../BuildToken'
 import createMutationGqlField from '../../createMutationGqlField'
 import createCollectionKeyInputHelpers from '../createCollectionKeyInputHelpers'
 import { getCollectionPatchType, getUpdateCollectionPayloadGqlType } from './createUpdateCollectionMutationFieldEntry'
+import getGqlOutputType from '../../type/getGqlOutputType'
 
 /**
  * Creates a update mutation which will update a single value from a collection
@@ -24,6 +25,7 @@ export default function createUpdateCollectionKeyMutationFieldEntry <TValue, TKe
   const inputHelpers = createCollectionKeyInputHelpers(buildToken, collectionKey)
   const patchFieldName = formatName.field(`${collection.type.name}-patch`)
   const { gqlType: patchGqlType, fromGqlInput: patchFromGqlInput } = getCollectionPatchType(buildToken, collection)
+  const { gqlType } = getGqlOutputType(buildToken, collection.type)
 
   return [formatName.field(name), createMutationGqlField<TValue>(buildToken, {
     name,
@@ -43,6 +45,6 @@ export default function createUpdateCollectionKeyMutationFieldEntry <TValue, TKe
     ],
     payloadType: getUpdateCollectionPayloadGqlType(buildToken, collection),
     execute: (context, input, resolveInfo) =>
-      collectionKey.update!(context, inputHelpers.getKeyFromInput(input), patchFromGqlInput(input[patchFieldName] as {}), resolveInfo),
+      collectionKey.update!(context, inputHelpers.getKeyFromInput(input), patchFromGqlInput(input[patchFieldName] as {}), resolveInfo, gqlType),
   })]
 }