- [x] calculate hasNextPage/hasPreviousPage in the same query as fetching the data
- [x] remove a source of N+1 queries relating to relations (foo.barsByBazId)

TODO:

- [ ] merge master
- [ ] try and remove additional parameters where possible (requires @calebmer's help)
- [ ] look into using CTEs to bypass performance issues, a la #396 

KNOWN ISSUES:

I've been informed by @cusxio that the following query returns null, but that including the `username` field fixes this - sounds like I need to add `username` to the list of implicit fields on the search endpoints:

```graphql
query {
  userByUsername(username:"cusx") {
    id
  }
}
```

Workaround:

```graphql
query {
  userByUsername(username:"cusx") {
    id
    username
  }
}
```


Fixes #219  
Fixes #265