@benjie 2017-02-04
I've just ran this against a large and relatively complex query for my own project;

master: 380ms, 435 SQL queries
this branch: 280ms, 9 SQL queries

That's with a local database (16GB RAM, quad core i7, virtually zero latency - ping says 0.06ms) - even if we factor in the network latency to the database to be just 1 millisecond we're looking at a significant speedup! (I just tested my servers and latency when unloaded comes in at an average about 1.2ms.)