@calebmer 2017-02-05
Those numbers are pretty sweet 👍

Maybe wait on merging into https://github.com/calebmer/postgraphql/pull/340? I’d like to get a chance to look at it first 😊