@benjie 2017-02-05
@calebmer Actually I think I'm going to merge this and #340 together - I was planning a two-stage thing (hence starting work on this separately) but as I progress I realise how much they are inter-dependent and not incorporating this into #340 may paint us into a corner preventing these advances from being implemented.