@benjie 2017-02-06
@valorize You're welcome to test it again; I'm currently aware that procedures returning sets are broken but most other things are working. I came up with a complex query for my dataset and then wrote the output of it on master to a JSON file, repeated on this branch and then diffed them. The aim is that the outputs are byte-for-byte identical.