@benjie 2017-02-06
I think all the read operations work (unless I just broke something...) and mutations trigger but they don't return data yet. I don't anticipate fixing that taking very long; but I'm done for the day. Hopefully tomorrow I'll get this into a state ready to use.