@benjie 2017-02-07
Most everything is working now; there's just one thing I need to fix which is reflective references - where a collection relates to itself; i.e.

```
Person {
  id
  name
  mother
  father
}


query {
  allPeople {
    nodes {
      name
      mother {
         name
      }
    }
  }
}
```

On it now...