@benjie 2017-02-07
Ah, there's an issue with procedures that return custom types... I gotta dash out but will fix later 👍