@valorize 2017-02-07
Hi Benji, I have a problem with authentication:
```
[{
"message": "Cannot read property 'selections' of null",
"locations": [{"line": 2, "column": 3}],
"path": ["authenticate"],
"stack": ["TypeError: Cannot read property 'selections' of null",
	"    at parseASTIntoFields (src/postgres/inventory/paginator/getSelectFragment.js:116:26)",
	"    at src/postgres/inventory/paginator/getSelectFragment.js:146:24", "    at Array.forEach (native)",
	"    at processFragment (src/postgres/inventory/paginator/getSelectFragment.js:145:42)",
	"    at parseASTIntoFields (src/postgres/inventory/paginator/getSelectFragment.js:133:5)",
	"    at src/postgres/inventory/paginator/getSelectFragment.js:118:9", "    at Array.forEach (native)",
	"    at parseASTIntoFields (src/postgres/inventory/paginator/getSelectFragment.js:116:38)",
	"    at src/postgres/inventory/paginator/getSelectFragment.js:15:7", "    at Array.forEach (native)"]
}]
```