@benjie 2017-02-08
Okay I've done some more major fixes.

<details>
<summary>We now pass the integration tests, except for one.</summary>

```
$ npm run test -- src/postgraphql/__tests__/*-test.*
 PASS  src/postgraphql/__tests__/postgraphqlIntegrationSchema-test.js
 FAIL  src/postgraphql/__tests__/postgraphqlIntegrationOperations-test.js (13.064s)
  ● operation procedure-query.graphql

    expect(value).toMatchSnapshot()

    Received value does not match stored snapshot 1.

    - Snapshot
    + Received

    @@ -9,42 +9,11 @@
           "b": "easy cheesy baked potatoes",
           "c": "RED",
           "d": null,
           "fooBar": 8,
         },
    -    "intSetQuery": Object {
    -      "edges": Array [
    -        Object {
    -          "cursor": "WyJuYXR1cmFsIiwxXQ==",
    -          "node": 1,
    -        },
    -        Object {
    -          "cursor": "WyJuYXR1cmFsIiwyXQ==",
    -          "node": 2,
    -        },
    -        Object {
    -          "cursor": "WyJuYXR1cmFsIiwzXQ==",
    -          "node": 3,
    -        },
    -        Object {
    -          "cursor": "WyJuYXR1cmFsIiw0XQ==",
    -          "node": 4,
    -        },
    -        Object {
    -          "cursor": "WyJuYXR1cmFsIiw1XQ==",
    -          "node": 5,
    -        },
    -        Object {
    -          "cursor": "WyJuYXR1cmFsIiw2XQ==",
    -          "node": null,
    -        },
    -        Object {
    -          "cursor": "WyJuYXR1cmFsIiw3XQ==",
    -          "node": 6,
    -        },
    -      ],
    -    },
    +    "intSetQuery": null,
         "noArgsQuery": 2,
         "tableQuery": Object {
           "__id": "WyJwb3N0cyIsNV0=",
           "authorId": 5,
           "headline": "Please, Don-Bot… look into your hard drive, and open your mercy file!",
    @@ -84,6 +53,9 @@
             },
           ],
         },
         "typesQuery": false,
       },
    +  "errors": Array [
    +    [GraphQLError: Expected integer. Not 'object'.],
    +  ],
     }

      at Object.<anonymous> (src/postgraphql/__tests__/postgraphqlIntegrationOperations-test.js:50:36)
      at step (node_modules/tslib/tslib.js:114:27)
      at Object.next (node_modules/tslib/tslib.js:95:57)
      at fulfilled (node_modules/tslib/tslib.js:85:62)
      at process._tickCallback (internal/process/next_tick.js:103:7)

 PASS  src/postgraphql/__tests__/postgraphql-test.js
 PASS  src/postgraphql/__tests__/withPostGraphQLContext-test.js

Snapshot Summary
 › 1 snapshot test failed in 1 test suite. Inspect your code changes or run with `npm test -- -u` to update them.

Test Suites: 1 failed, 3 passed, 4 total
Tests:       1 failed, 42 passed, 43 total
Snapshots:   1 failed, 18 passed, 19 total
Time:        15.544s
Ran all test suites matching /src/postgraphql/__tests__/postgraphql-test.js|src/postgraphql/__tests__/postgraphqlIntegrationOperations-test.js|src/postgraphql/__tests__/postgraphqlIntegrationSchema-test.js|src/postgraphql/__tests__/withPostGraphQLContext-test.js/.
```
</details>

The one that's failing is the `intSetQuery()` custom procedure which returns a set of int

https://github.com/calebmer/postgraphql/blob/39cff1d123c282667e339064a9a5be9538e330a6/src/postgraphql/__tests__/fixtures/queries/procedure-query.graphql#L10
https://github.com/calebmer/postgraphql/blob/39cff1d123c282667e339064a9a5be9538e330a6/examples/kitchen-sink/schema.sql#L163

I've dug around a bunch but I can't figure out how to make specifically this one case return the node directly; at the moment it's getting returned a value that's like `{"__local_1__": 1}` rather than simply `1`. This should be solveable by putting the relevant `sqlExpression` in the right place or merely hacking the resolver to return the one and only property of the object; but I can't quite figure it out.

I also guarantee NOTHING when it comes to `classicIds`; the rewriting of fieldName messes with my assumptions and I'll need @calebmer's help figuring out how to correctly account for this. I've made a start, enough to pass the tests above, but I'm not confident in it.

I'm going to be testing this tomorrow on my own stuff, if you want to give it a go you can `npm install calebmer/postgraphql#ignore/bg2` but beware that branch changes frequently and is constantly rebased so DO NOT RELY ON THE COMMIT HASHES!

I've also noticed that because computed columns on rows you update via an SQL function use the value of the row BEFORE the update rather than after. E.g. this would be a problem:

```
mutation {
  favoritePost(id: $id) {
    post {
      isFavorite
    }
  }
}
```